Resolve #21 and #22
We decided to drop all spots that have a low library size and were on the edge, on either the whole genome or the targeted sequencing objects. This ensures that the spots remaining are the same for both analyses, which will simplify comparing them against each other.

Fix the edge distance calculation, related to #22. Note how the edge only really affects 2 spots in one sample, which didn't really affects us downstream, so we decided against manually fixing the edge for them. Also, the edge distance calculation is simplistic in the sense that we are calculating the distance to the closest edge on either the same row or column, but not in diagonals.

Co-authored-by: Sang Ho Kwon <skwon17@jh.edu>